requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector-python==8.0.28
packaging==21.3
protobuf==3.19.4
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
QtAwesome==1.1.1
QtPy==2.0.1
requests==2.27.1
urllib3==1.26.9
